+++ tokens/components/button.token.ts
@@ -1,0 +1,49 @@
+import { primaryTokens } from "../primary/primary.token"
+/**
+ * This file include the component token for buntton
+ */
+
+
+
+
+
+
+
+
+const tokensBase = {
+    "button-border-radius": primaryTokens["radius-nl"],
+    "button-border-gradient": primaryTokens["gradient-semantic2define-1"]
+}
+const primaryBase = {
+    "button-primary-text": primaryTokens["color-white"],
+    "button-primary-icon": primaryTokens["color-white"]
+
+}
+const primaryDefault = {
+    "button-primary-background": primaryTokens["color-primary-medium"],
+
+}
+const primaryHover = {
+    "button-primary-background": primaryTokens["color-primary-high"],
+}
+const primaryDisabled = {
+    "button-primary-background": primaryTokens["color-tertiary-low"],
+}
+
+
+const tokensDanger = {
+
+}
+
+
+
+
+
+const tokens = {
+    base: tokensBase,
+    primary: {
+        base: primaryBase,
+        default: primaryDefault,
+        primaryHover: primaryHover
+    },
+}
