+++ tokens/primary/backgrounds.tokens.ts
@@ -1,0 +1,9 @@
+import { primaryColorTokens } from "./colors.tokens"
+
+export const primaryBackgroundTokens = {
+
+    "background-primary": primaryColorTokens["color-tertiary-high"],
+    "background-secondary": primaryColorTokens["color-tertiary-medium"],
+    "background-tertiary": primaryColorTokens["color-tertiary-low"],
+
+}
