+++ tokens/primary/colors.tokens.ts
@@ -1,0 +1,30 @@
+export const primaryColorTokens = {
+    "color-primary-high": "#684CF4",
+    "color-primary-medium": "#8770F9",
+    "color-primary-low": "#A897FB",
+
+    "color-secondary-high": "#15FC80",
+    "color-secondary-medium": "#6FFFB2",
+    "color-secondary-low": "#DEFFED",
+
+    "color-tertiary-high": "#050507",
+    "color-tertiary-medium": "#1D1E22",
+    "color-tertiary-low": "#39393D",
+    "color-tertiary-lower": "#969696",
+    "color-tertiary-lowest": "#BFBFBF",
+
+    "color-danger-high": "#EB5557",
+
+    "color-warning-high": "#FFA45B",
+
+    "color-white": "#F1F1F1",
+
+
+
+
+
+    "color-tertiary": "rgba(134, 112, 249, 1)", // Should be removed
+    "radius-nl": "4px",
+
+    "gradient-semantic2define-1": "linear-gradient(180deg, #6B6B6B 0%, rgba(153, 153, 153, 0) 100%)",
+}
