+++ tokens/primary/primary.token.ts
@@ -1,0 +1,7 @@
+import { primaryBackgroundTokens, } from "./backgrounds.tokens"
+import { primaryColorTokens } from "./colors.tokens"
+
+export const primaryTokens = {
+    ...primaryColorTokens,
+    ...primaryBackgroundTokens
+}
